move 'tsql::parser::types::TableExtra' to 'tsql::types::TableExtra'
